Revert "Разделение одного маппера, на много мелких мапперов"
This reverts commit 87e3e78f82e733b8e8987b9237ddc5d512be20c3.
